requirements.txt
transformers>=4.44.0
accelerate>=1.0.0
pillow>=10.0.0
